[Refactor] - Extract JsonRpcHttpOptions to a new class  (#6478)
