+++ besu/src/test/java/org/hyperledger/besu/cli/BesuCommandTest.java
@@ -28,9 +28,6 @@ import static org.hyperledger.besu.cli.config.NetworkName.MAINNET;
 import static org.hyperledger.besu.cli.config.NetworkName.MORDOR;
 import static org.hyperledger.besu.cli.config.NetworkName.SEPOLIA;
 import static org.hyperledger.besu.ethereum.api.jsonrpc.RpcApis.ENGINE;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.RpcApis.ETH;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.RpcApis.NET;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.RpcApis.PERM;
 import static org.hyperledger.besu.ethereum.p2p.config.DefaultDiscoveryConfiguration.GOERLI_BOOTSTRAP_NODES;
 import static org.hyperledger.besu.ethereum.p2p.config.DefaultDiscoveryConfiguration.GOERLI_DISCOVERY_URL;
 import static org.hyperledger.besu.ethereum.p2p.config.DefaultDiscoveryConfiguration.MAINNET_BOOTSTRAP_NODES;
@@ -45,7 +42,6 @@ import static org.mockito.ArgumentMatchers.isNotNull;
 import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
@@ -64,10 +60,7 @@ import org.hyperledger.besu.ethereum.api.ImmutableApiConfiguration;
 import org.hyperledger.besu.ethereum.api.graphql.GraphQLConfiguration;
 import org.hyperledger.besu.ethereum.api.handlers.TimeoutOptions;
 import org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcConfiguration;
-import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
-import org.hyperledger.besu.ethereum.api.jsonrpc.authentication.JwtAlgorithm;
 import org.hyperledger.besu.ethereum.api.jsonrpc.websocket.WebSocketConfiguration;
-import org.hyperledger.besu.ethereum.api.tls.TlsConfiguration;
 import org.hyperledger.besu.ethereum.core.MiningParameters;
 import org.hyperledger.besu.ethereum.core.PrivacyParameters;
 import org.hyperledger.besu.ethereum.eth.sync.SyncMode;
@@ -87,7 +80,6 @@ import org.hyperledger.besu.nat.NatMethod;
 import org.hyperledger.besu.pki.config.PkiKeyStoreConfiguration;
 import org.hyperledger.besu.plugin.data.EnodeURL;
 import org.hyperledger.besu.plugin.services.privacy.PrivateMarkerTransactionFactory;
-import org.hyperledger.besu.plugin.services.rpc.PluginRpcRequest;
 import org.hyperledger.besu.util.number.Fraction;
 import org.hyperledger.besu.util.number.Percentage;
 import org.hyperledger.besu.util.number.PositiveNumber;
@@ -96,7 +88,6 @@ import org.hyperledger.besu.util.platform.PlatformDetector;
 import java.io.File;
 import java.io.IOException;
 import java.math.BigInteger;
-import java.net.ServerSocket;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
@@ -110,7 +101,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -1546,37 +1536,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpMaxBatchSizeOptionMustBeUsed() {
-    final int rpcHttpMaxBatchSize = 1;
-    parseCommand("--rpc-http-max-batch-size", Integer.toString(rpcHttpMaxBatchSize));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getMaxBatchSize())
-        .isEqualTo(rpcHttpMaxBatchSize);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpMaxRequestContentLengthOptionMustBeUsed() {
-    final int rpcHttpMaxRequestContentLength = 1;
-    parseCommand(
-        "--rpc-http-max-request-content-length", Long.toString(rpcHttpMaxRequestContentLength));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getMaxRequestContentLength())
-        .isEqualTo(rpcHttpMaxRequestContentLength);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
   public void maxpeersSet_p2pPeerLowerBoundSet() {
 
     final int maxPeers = 123;
@@ -1998,19 +1957,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcHttpEnabledPropertyMustBeUsed() {
-    parseCommand("--rpc-http-enabled");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().isEnabled()).isTrue();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
   public void graphQLHttpEnabledPropertyDefaultIsFalse() {
     parseCommand();
 
@@ -2037,22 +1983,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcApisPropertyMustBeUsed() {
-    parseCommand("--rpc-http-api", "ETH,NET,PERM", "--rpc-http-enabled");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-    verify(mockLogger)
-        .warn("Permissions are disabled. Cannot enable PERM APIs when not using Permissions.");
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getRpcApis())
-        .containsExactlyInAnyOrder(ETH.name(), NET.name(), PERM.name());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
   public void rpcApisSupportsEngine() {
     parseCommand("--rpc-http-api", "ENGINE", "--rpc-http-enabled");
 
@@ -2061,38 +1991,6 @@ public class BesuCommandTest extends CommandTestAbstract {
 
     assertThat(jsonRpcConfigArgumentCaptor.getValue().getRpcApis())
         .containsExactlyInAnyOrder(ENGINE.name());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcApisPropertyIgnoresDuplicatesAndMustBeUsed() {
-    parseCommand("--rpc-http-api", "ETH,NET,NET", "--rpc-http-enabled");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getRpcApis())
-        .containsExactlyInAnyOrder(ETH.name(), NET.name());
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcApiNoAuthMethodsIgnoresDuplicatesAndMustBeUsed() {
-    parseCommand(
-        "--rpc-http-api-methods-no-auth",
-        "admin_peers, admin_peers, eth_getWork",
-        "--rpc-http-enabled");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getNoAuthRpcApis())
-        .containsExactlyInAnyOrder(
-            RpcMethod.ADMIN_PEERS.getMethodName(), RpcMethod.ETH_GET_WORK.getMethodName());
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
     assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
@@ -2127,69 +2025,6 @@ public class BesuCommandTest extends CommandTestAbstract {
         "/tmp/fakeKey.hex");
     verify(mockRunnerBuilder).engineJsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
     assertThat(jsonRpcConfigArgumentCaptor.getValue().isAuthenticationEnabled()).isFalse();
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpNoAuthApiMethodsCannotBeInvalid() {
-    parseCommand("--rpc-http-enabled", "--rpc-http-api-method-no-auth", "invalid");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "Invalid value for option '--rpc-http-api-methods-no-auth', options must be valid RPC methods");
-  }
-
-  @Test
-  public void rpcHttpOptionsRequiresServiceToBeEnabled() {
-    parseCommand(
-        "--rpc-http-api",
-        "ETH,NET",
-        "--rpc-http-host",
-        "0.0.0.0",
-        "--rpc-http-port",
-        "1234",
-        "--rpc-http-cors-origins",
-        "all",
-        "--rpc-http-max-active-connections",
-        "88");
-
-    verifyOptionsConstraintLoggerCall(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        "--rpc-http-port",
-        "--rpc-http-cors-origins",
-        "--rpc-http-api",
-        "--rpc-http-max-active-connections");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpOptionsRequiresServiceToBeEnabledToml() throws IOException {
-    final Path toml =
-        createTempFile(
-            "toml",
-            "rpc-http-api=[\"ETH\",\"NET\"]\n"
-                + "rpc-http-host=\"0.0.0.0\"\n"
-                + "rpc-http-port=1234\n"
-                + "rpc-http-cors-origins=[\"all\"]\n"
-                + "rpc-http-max-active-connections=88");
-
-    parseCommand("--config-file", toml.toString());
-
-    verifyOptionsConstraintLoggerCall(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        "--rpc-http-port",
-        "--rpc-http-cors-origins",
-        "--rpc-http-api",
-        "--rpc-http-max-active-connections");
-
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
     assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
   }
@@ -2277,498 +2112,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void rpcApisPropertyWithInvalidEntryMustDisplayError() {
-    parseCommand("--rpc-http-api", "BOB");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    // PicoCLI uses longest option name for message when option has multiple names, so here plural.
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Invalid value for option '--rpc-http-api': invalid entries found [BOB]");
-  }
-
-  @Test
-  public void rpcApisPropertyWithPluginNamespaceAreValid() {
-
-    rpcEndpointServiceImpl.registerRPCEndpoint(
-        "bob", "method", (Function<PluginRpcRequest, Object>) request -> "nothing");
-
-    parseCommand("--rpc-http-api", "BOB");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getRpcApis())
-        .containsExactlyInAnyOrder("BOB");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpHostAndPortOptionsMustBeUsed() {
-
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    parseCommand(
-        "--rpc-http-enabled", "--rpc-http-host", host, "--rpc-http-port", String.valueOf(port));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpHostMayBeLocalhost() {
-
-    final String host = "localhost";
-    parseCommand("--rpc-http-enabled", "--rpc-http-host", host);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpHostMayBeIPv6() {
-
-    final String host = "2600:DB8::8545";
-    parseCommand("--rpc-http-enabled", "--rpc-http-host", host);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpMaxActiveConnectionsPropertyMustBeUsed() {
-    final int maxConnections = 99;
-    parseCommand("--rpc-http-max-active-connections", String.valueOf(maxConnections));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getMaxActiveConnections())
-        .isEqualTo(maxConnections);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsRequiresRpcHttpEnabled() {
-    parseCommand("--rpc-http-tls-enabled");
-
-    verifyOptionsConstraintLoggerCall("--rpc-http-enabled", "--rpc-http-tls-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsRequiresRpcHttpEnabledToml() throws IOException {
-    final Path toml = createTempFile("toml", "rpc-http-tls-enabled=true\n");
-
-    parseCommand("--config-file", toml.toString());
-
-    verifyOptionsConstraintLoggerCall("--rpc-http-enabled", "--rpc-http-tls-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsWithoutKeystoreReportsError() {
-    parseCommand("--rpc-http-enabled", "--rpc-http-tls-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Keystore file is required when TLS is enabled for JSON-RPC HTTP endpoint");
-  }
-
-  @Test
-  public void rpcHttpTlsWithoutPasswordfileReportsError() {
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        "/tmp/test.p12");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "File containing password to unlock keystore is required when TLS is enabled for JSON-RPC HTTP endpoint");
-  }
-
-  @Test
-  public void rpcHttpTlsKeystoreAndPasswordMustBeUsed() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration.isPresent()).isTrue();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().isEmpty()).isTrue();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsClientAuthWithoutKnownFileReportsError() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-client-auth-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "Known-clients file must be specified or CA clients must be enabled when TLS client authentication is enabled for JSON-RPC HTTP endpoint");
-  }
-
-  @Test
-  public void rpcHttpTlsClientAuthWithKnownClientFile() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    final String knownClientFile = "/tmp/knownClientFile";
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-client-auth-enabled",
-        "--rpc-http-tls-known-clients-file",
-        knownClientFile);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration.isPresent()).isTrue();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().isPresent()).isTrue();
-    assertThat(
-            tlsConfiguration.get().getClientAuthConfiguration().get().getKnownClientsFile().get())
-        .isEqualTo(Path.of(knownClientFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().get().isCaClientsEnabled())
-        .isFalse();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsClientAuthWithCAClient() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-client-auth-enabled",
-        "--rpc-http-tls-ca-clients-enabled");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration.isPresent()).isTrue();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().isPresent()).isTrue();
-    assertThat(
-            tlsConfiguration
-                .get()
-                .getClientAuthConfiguration()
-                .get()
-                .getKnownClientsFile()
-                .isEmpty())
-        .isTrue();
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().get().isCaClientsEnabled())
-        .isTrue();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsClientAuthWithCAClientAndKnownClientFile() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    final String knownClientFile = "/tmp/knownClientFile";
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-client-auth-enabled",
-        "--rpc-http-tls-ca-clients-enabled",
-        "--rpc-http-tls-known-clients-file",
-        knownClientFile);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration.isPresent()).isTrue();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().isPresent()).isTrue();
-    assertThat(
-            tlsConfiguration.get().getClientAuthConfiguration().get().getKnownClientsFile().get())
-        .isEqualTo(Path.of(knownClientFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration().get().isCaClientsEnabled())
-        .isTrue();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsCheckDefaultProtocolsAndCipherSuites() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration).isPresent();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration()).isEmpty();
-    assertThat(tlsConfiguration.get().getCipherSuites().get()).isEmpty();
-    assertThat(tlsConfiguration.get().getSecureTransportProtocols().get())
-        .containsExactly("TLSv1.3", "TLSv1.2");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsCheckInvalidProtocols() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    final String protocol = "TLsv1.4";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-protocols",
-        protocol);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).contains("No valid TLS protocols specified");
-  }
-
-  @Test
-  public void rpcHttpTlsCheckInvalidCipherSuites() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    final String cipherSuites = "Invalid";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-cipher-suites",
-        cipherSuites);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Invalid TLS cipher suite specified " + cipherSuites);
-  }
-
-  @Test
-  public void rpcHttpTlsCheckValidProtocolsAndCipherSuites() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String keystoreFile = "/tmp/test.p12";
-    final String keystorePasswordFile = "/tmp/test.txt";
-    final String protocols = "TLSv1.3,TLSv1.2";
-    final String cipherSuites =
-        "TLS_AES_256_GCM_SHA384,TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-enabled",
-        "--rpc-http-tls-keystore-file",
-        keystoreFile,
-        "--rpc-http-tls-keystore-password-file",
-        keystorePasswordFile,
-        "--rpc-http-tls-protocols",
-        protocols,
-        "--rpc-http-tls-cipher-suites",
-        cipherSuites);
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getHost()).isEqualTo(host);
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    final Optional<TlsConfiguration> tlsConfiguration =
-        jsonRpcConfigArgumentCaptor.getValue().getTlsConfiguration();
-    assertThat(tlsConfiguration).isPresent();
-    assertThat(tlsConfiguration.get().getKeyStorePath()).isEqualTo(Path.of(keystoreFile));
-    assertThat(tlsConfiguration.get().getClientAuthConfiguration()).isEmpty();
-    assertThat(tlsConfiguration.get().getCipherSuites().get())
-        .containsExactlyInAnyOrder(
-            "TLS_AES_256_GCM_SHA384",
-            "TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384",
-            "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256");
-    assertThat(tlsConfiguration.get().getSecureTransportProtocols().get())
-        .containsExactlyInAnyOrder("TLSv1.2", "TLSv1.3");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpTlsWarnIfCipherSuitesSpecifiedWithoutTls() {
-    final String host = "1.2.3.4";
-    final int port = 1234;
-    final String cipherSuites = "Invalid";
-
-    parseCommand(
-        "--rpc-http-enabled",
-        "--engine-rpc-enabled",
-        "--rpc-http-host",
-        host,
-        "--rpc-http-port",
-        String.valueOf(port),
-        "--rpc-http-tls-cipher-suite",
-        cipherSuites);
-    verify(
-            mockLogger,
-            times(2)) // this is verified for both the full suite of apis, and the engine group.
-        .warn(
-            "{} has been ignored because {} was not defined on the command line.",
-            "--rpc-http-tls-cipher-suite",
-            "--rpc-http-tls-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
   public void graphQLHttpHostAndPortOptionsMustBeUsed() {
 
     final String host = "1.2.3.4";
@@ -2818,169 +2161,6 @@ public class BesuCommandTest extends CommandTestAbstract {
 
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
     assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsTwoDomainsMustBuildListWithBothDomains() {
-    final String[] origins = {"http://domain1.com", "https://domain2.com"};
-    parseCommand("--rpc-http-enabled", "--rpc-http-cors-origins", String.join(",", origins));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getCorsAllowedDomains().toArray())
-        .isEqualTo(origins);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsDoubleCommaFilteredOut() {
-    final String[] origins = {"http://domain1.com", "https://domain2.com"};
-    parseCommand("--rpc-http-enabled", "--rpc-http-cors-origins", String.join(",,", origins));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getCorsAllowedDomains().toArray())
-        .isEqualTo(origins);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsWithWildcardMustBuildListWithWildcard() {
-    final String[] origins = {"*"};
-    parseCommand("--rpc-http-enabled", "--rpc-http-cors-origins", String.join(",", origins));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getCorsAllowedDomains().toArray())
-        .isEqualTo(origins);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsWithAllMustBuildListWithWildcard() {
-    parseCommand("--rpc-http-enabled", "--rpc-http-cors-origins", "all");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getCorsAllowedDomains()).containsExactly("*");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsWithNoneMustBuildEmptyList() {
-    final String[] origins = {"none"};
-    parseCommand("--rpc-http-enabled", "--rpc-http-cors-origins", String.join(",", origins));
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getCorsAllowedDomains()).isEmpty();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsNoneWithAnotherDomainMustFail() {
-    final String[] origins = {"http://domain1.com", "none"};
-    parseCommand("--rpc-http-cors-origins", String.join(",", origins));
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Value 'none' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsNoneWithAnotherDomainMustFailNoneFirst() {
-    final String[] origins = {"none", "http://domain1.com"};
-    parseCommand("--rpc-http-cors-origins", String.join(",", origins));
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Value 'none' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsAllWithAnotherDomainMustFail() {
-    parseCommand("--rpc-http-cors-origins=http://domain1.com,all");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Values '*' or 'all' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsAllWithAnotherDomainMustFailAsFlags() {
-    parseCommand("--rpc-http-cors-origins=http://domain1.com", "--rpc-http-cors-origins=all");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Values '*' or 'all' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsWildcardWithAnotherDomainMustFail() {
-    parseCommand("--rpc-http-cors-origins=http://domain1.com,*");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Values '*' or 'all' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsWildcardWithAnotherDomainMustFailAsFlags() {
-    parseCommand("--rpc-http-cors-origins=http://domain1.com", "--rpc-http-cors-origins=*");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Values '*' or 'all' can't be used with other domains");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsInvalidRegexShouldFail() {
-    final String[] origins = {"**"};
-    parseCommand("--rpc-http-cors-origins", String.join(",", origins));
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Domain values result in invalid regex pattern");
-  }
-
-  @Test
-  public void rpcHttpCorsOriginsEmptyValueFails() {
-    parseCommand("--rpc-http-cors-origins=");
-
-    Mockito.verifyNoInteractions(mockRunnerBuilder);
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Domain cannot be empty string or null string.");
   }
 
   /** test deprecated CLI option * */
@@ -3165,19 +2345,6 @@ public class BesuCommandTest extends CommandTestAbstract {
     assertThat(commandOutput.toString(UTF_8)).isEmpty();
     assertThat(commandErrorOutput.toString(UTF_8))
         .contains("Hostname cannot be empty string or null string.");
-  }
-
-  @Test
-  public void rpcWsRpcEnabledPropertyDefaultIsFalse() {
-    parseCommand();
-
-    verify(mockRunnerBuilder).webSocketConfiguration(wsRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(wsRpcConfigArgumentCaptor.getValue().isEnabled()).isFalse();
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
   }
 
   @Test
@@ -4148,40 +3315,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void httpAuthenticationAlgorithIsConfigured() {
-    parseCommand("--rpc-http-authentication-jwt-algorithm", "ES256");
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getAuthenticationAlgorithm())
-        .isEqualTo(JwtAlgorithm.ES256);
-  }
-
-  @Test
-  public void httpAuthenticationPublicKeyIsConfigured() throws IOException {
-    final Path publicKey = Files.createTempFile("public_key", "");
-    parseCommand("--rpc-http-authentication-jwt-public-key-file", publicKey.toString());
-
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getAuthenticationPublicKeyFile().getPath())
-        .isEqualTo(publicKey.toString());
-  }
-
-  @Test
-  public void httpAuthenticationWithoutRequiredConfiguredOptionsMustFail() {
-    parseCommand("--rpc-http-enabled", "--rpc-http-authentication-enabled");
-
-    verifyNoInteractions(mockRunnerBuilder);
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "Unable to authenticate JSON-RPC HTTP endpoint without a supplied credentials file or authentication public key file");
-  }
-
-  @Test
   public void privHttpApisWithPrivacyDisabledLogsWarning() {
     parseCommand("--privacy-enabled=false", "--rpc-http-api", "PRIV", "--rpc-http-enabled");
 
@@ -4413,29 +3546,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   }
 
   @Test
-  public void assertThatCheckPortClashAcceptsAsExpected() throws Exception {
-    // use WS port for HTTP
-    final int port = 8546;
-    parseCommand("--rpc-http-enabled", "--rpc-http-port", String.valueOf(port));
-    verify(mockRunnerBuilder).jsonRpcConfiguration(jsonRpcConfigArgumentCaptor.capture());
-    verify(mockRunnerBuilder).build();
-    assertThat(jsonRpcConfigArgumentCaptor.getValue().getPort()).isEqualTo(port);
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
-  }
-
-  @Test
-  public void assertThatCheckPortClashRejectsAsExpected() throws Exception {
-    // use WS port for HTTP
-    final int port = 8546;
-    parseCommand("--rpc-http-enabled", "--rpc-http-port", String.valueOf(port), "--rpc-ws-enabled");
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains(
-            "Port number '8546' has been specified multiple times. Please review the supplied configuration.");
-  }
-
-  @Test
   public void assertThatCheckPortClashRejectsAsExpectedForEngineApi() throws Exception {
     // use WS port for HTTP
     final int port = 8545;
@@ -4636,19 +3746,6 @@ public class BesuCommandTest extends CommandTestAbstract {
   public void logWarnIfFastSyncMinPeersUsedWithFullSync() {
     parseCommand("--sync-mode", "FULL", "--fast-sync-min-peers", "1");
     verify(mockLogger).warn("--sync-min-peers is ignored in FULL sync-mode");
-  }
-
-  @Test
-  public void portInUseReportsError() throws IOException {
-    final ServerSocket serverSocket = new ServerSocket(8545);
-
-    parseCommandWithPortCheck("--rpc-http-enabled");
-
-    assertThat(commandOutput.toString(UTF_8)).isEmpty();
-    assertThat(commandErrorOutput.toString(UTF_8))
-        .contains("Port(s) '[8545]' already in use. Check for other processes using the port(s).");
-
-    serverSocket.close();
   }
 
   @Test
